+++ src/Helper/DateUtcHelper.php
@@ -1,34 +1,25 @@
 <?php
 namespace DalPraS\OAuth2\Client\Helper;
 
-class DateUtcHelper {
-
-    /**
-     * @var \DateTimeZone
-     */
-    private $utcTimeZone;
-
-    /**
-     * @param \DateTime $dateTime
-     */
-    public function __construct() {
-        $this->utcTimeZone = new \DateTimeZone('UTC');
-    }
-
+class DateUtcHelper 
+{
     /**
      * Convert DateTime in UTC timezone string format Y-m-d\TH:i:s\Z
      *
      * @param \DateTime $dateTime Local datetime
      * @return string
      */
-    public function date2utc(\DateTime $dateTime) : string {
-        return $dateTime->setTimezone($this->utcTimeZone)->format('Y-m-d\TH:i:s\Z');
+    public static function date2utc(\DateTime $dateTime) : string {
+        return $dateTime->setTimezone(new \DateTimeZone('UTC'))->format('Y-m-d\TH:i:s\Z');
     }
 
     /**
      * Convert string UTC datetime in current datetime
+     * 
+     * @param string $utcTime String representing the time
+     * @return \DateTime
      */
-    public function utc2date(string $utcTime) : \DateTime {
+    public static function utc2date(string $utcTime) : \DateTime {
         return (new \DateTime())->createFromFormat('Y-m-d\TH:i:s\Z', $utcTime);
     }
 }