+++ src/Resources/CoOrganizer.php
@@ -3,7 +3,6 @@
 namespace DalPraS\OAuth2\Client\Resources;
 
 use DalPraS\OAuth2\Client\ResultSet\SimpleResultSet;
-use DalPraS\OAuth2\Client\ResultSet\ResultSetInterface;
 
 class CoOrganizer extends AuthenticatedResourceAbstract
 {
@@ -16,7 +15,7 @@ class CoOrganizer extends AuthenticatedResourceAbstract
      *
      * @link https://developer.goto.com/GoToWebinarV2/#operation/getCoorganizers
      */
-    public function getCoOrganizers(string $webinarKey): ResultSetInterface
+    public function getCoOrganizers(string $webinarKey): SimpleResultSet
     {
         $url = $this->getRequestUrl('/organizers/{organizerKey}/webinars/{webinarKey}/coorganizers', [
             'webinarKey' => $webinarKey,
@@ -35,7 +34,7 @@ class CoOrganizer extends AuthenticatedResourceAbstract
      *
      * @link https://developer.goto.com/GoToWebinarV2/#operation/createCoorganizers
      */
-    public function createCoOrganizers(string $webinarKey, array $body): ResultSetInterface
+    public function createCoOrganizers(string $webinarKey, array $body): SimpleResultSet
     {
         $url = $this->getRequestUrl('/organizers/{organizerKey}/webinars/{webinarKey}/coorganizers', [
             'webinarKey' => $webinarKey
@@ -57,7 +56,7 @@ class CoOrganizer extends AuthenticatedResourceAbstract
      *
      * @link https://developer.goto.com/GoToWebinarV2/#operation/deleteCoorganizer
      */
-    public function deleteCoOrganizer(string $webinarKey, string $coOrganizerKey, bool $external = false): ResultSetInterface
+    public function deleteCoOrganizer(string $webinarKey, string $coOrganizerKey, bool $external = false): SimpleResultSet
     {
         $url = $this->getRequestUrl('/organizers/{organizerKey}/webinars/{webinarKey}/coorganizers/{coorganizerKey}', [
             'webinarKey' => $webinarKey,
@@ -71,17 +70,39 @@ class CoOrganizer extends AuthenticatedResourceAbstract
     }
 
     /**
-     * Resend invitation
+     * Resends an invitation email to the specified co-organizer. For external co-organizers 
+     * (individuals who do not have a shared GoToWebinar account), set the URL parameter 'external' = true.
+     * 
+     * https://api.getgo.com/G2W/rest/v2/organizers/{organizerKey}/webinars/{webinarKey}/coorganizers/{coorganizerKey}/resendInvitation
      *
+     * @deprecated use resendCoOrganizerInvitation
+     * 
      * @param string $webinarKey
      * @param string $coOrganizerKey
      * @param bool $external
-     * @return array|null
      * @throws \League\OAuth2\Client\Provider\Exception\IdentityProviderException
      *
      * @link https://developer.goto.com/GoToWebinarV2/#operation/resendCoorganizerInvitation
      */
-    public function resendInvitation(string $webinarKey, string $coOrganizerKey, bool $external = false): ResultSetInterface
+    public function resendInvitation(string $webinarKey, string $coOrganizerKey, bool $external = false): SimpleResultSet
+    {
+        return $this->resendCoOrganizerInvitation($webinarKey, $coOrganizerKey, $external);
+    }
+
+    /**
+     * Resends an invitation email to the specified co-organizer. For external co-organizers 
+     * (individuals who do not have a shared GoToWebinar account), set the URL parameter 'external' = true.
+     * 
+     * https://api.getgo.com/G2W/rest/v2/organizers/{organizerKey}/webinars/{webinarKey}/coorganizers/{coorganizerKey}/resendInvitation
+     * 
+     * @param string $webinarKey
+     * @param string $coOrganizerKey
+     * @param bool $external
+     * @throws \League\OAuth2\Client\Provider\Exception\IdentityProviderException
+     *
+     * @link https://developer.goto.com/GoToWebinarV2/#operation/resendCoorganizerInvitation
+     */
+    public function resendCoOrganizerInvitation(string $webinarKey, string $coOrganizerKey, bool $external = false): SimpleResultSet
     {
         $url = $this->getRequestUrl('/organizers/{organizerKey}/webinars/{webinarKey}/coorganizers/{coorganizerKey}/resendInvitation', [
             'webinarKey' => $webinarKey,