+++ src/Resources/Registrant.php
@@ -3,7 +3,6 @@
 namespace DalPraS\OAuth2\Client\Resources;
 
 use DalPraS\OAuth2\Client\ResultSet\SimpleResultSet;
-use DalPraS\OAuth2\Client\ResultSet\ResultSetInterface;
 
 class Registrant extends \DalPraS\OAuth2\Client\Resources\AuthenticatedResourceAbstract
 {
@@ -15,7 +14,7 @@ class Registrant extends \DalPraS\OAuth2\Client\Resources\AuthenticatedResourceA
      *
      * @param string $webinarKey
      */
-    public function getRegistrants(string $webinarKey): ResultSetInterface
+    public function getRegistrants(string $webinarKey): SimpleResultSet
     {
         $url = $this->getRequestUrl('/organizers/{organizerKey}/webinars/{webinarKey}/registrants', [
             'webinarKey' => $webinarKey,
@@ -33,7 +32,7 @@ class Registrant extends \DalPraS\OAuth2\Client\Resources\AuthenticatedResourceA
      * @param string $webinarKey
      * @param string $registrantKey
      */
-    public function getRegistrant(string $webinarKey, string $registrantKey): ResultSetInterface
+    public function getRegistrant(string $webinarKey, string $registrantKey): SimpleResultSet
     {
         $url = $this->getRequestUrl('/organizers/{organizerKey}/webinars/{webinarKey}/registrants/{registrantKey}', [
             'webinarKey' => $webinarKey,
@@ -52,7 +51,7 @@ class Registrant extends \DalPraS\OAuth2\Client\Resources\AuthenticatedResourceA
      * @param string $email
      * @return array|NULL
      */
-    public function getRegistrantByEmail(string $webinarKey, string $email): ResultSetInterface
+    public function getRegistrantByEmail(string $webinarKey, string $email): SimpleResultSet
     {
         $registrants = $this->getRegistrants($webinarKey);
         foreach ($registrants as $registrant) {
@@ -64,19 +63,25 @@ class Registrant extends \DalPraS\OAuth2\Client\Resources\AuthenticatedResourceA
     }
 
     /**
-     * Subscribe a registrant for a webinar.
+     * Register an attendee for a scheduled webinar. Requires the organizer and a scheduled webinar.
+     * Please note that you must provide all required fields including custom fields defined during the webinar creation. 
+     * Use the API call 'Get registration fields' to get a list of all fields, if they are required, and their possible values. 
+     * 
      * https://api.getgo.com/G2W/rest/v2/organizers/{organizerKey}/webinars/{webinarKey}/registrants
      *
      * @link https://developer.goto.com/GoToWebinarV2#operation/createRegistrant
      *
      * @param string $webinarKey
-     * @param array $body
-     * @return array
+     * @param array $body 
+     * @param bool $resendConfirmation
+     *      Indicates whether the confirmation email should be resent when a registrant is re-registered
      */
-    public function createRegistrant(string $webinarKey, array $body): ResultSetInterface
+    public function createRegistrant(string $webinarKey, array $body, bool $resendConfirmation = false): SimpleResultSet
     {
         $url = $this->getRequestUrl('/organizers/{organizerKey}/webinars/{webinarKey}/registrants', [
             'webinarKey' => $webinarKey,
+        ], [
+            'resendConfirmation' => $resendConfirmation ? 'true' : 'false'
         ]);
         $request  = $this->provider->getAuthenticatedRequest('POST', $url, $this->accessToken, [
             'headers' => [
@@ -88,6 +93,20 @@ class Registrant extends \DalPraS\OAuth2\Client\Resources\AuthenticatedResourceA
     }
 
     /**
+     * Retrieve required, optional registration, and custom questions for a specified webinar.
+     * 
+     * https://api.getgo.com/G2W/rest/v2/organizers/{organizerKey}/webinars/{webinarKey}/registrants/fields
+     * 
+     * @link https://developer.goto.com/GoToWebinarV2#operation/getRegistrationFields
+     */
+    public function getRegistrationFields(): SimpleResultSet 
+    {
+        $url = $this->getRequestUrl('/organizers/{organizerKey}/webinars/{webinarKey}/registrants/fields');
+        $request  = $this->provider->getAuthenticatedRequest('GET', $url, $this->accessToken);
+        return new SimpleResultSet($this->provider->getParsedResponse($request));
+    }
+    
+    /**
      * Unsubscribe a registrant from a webinar.
      * https://api.getgo.com/G2W/rest/v2/organizers/{organizerKey}/webinars/{webinarKey}/registrants/{registrantKey}
      *
@@ -97,7 +116,7 @@ class Registrant extends \DalPraS\OAuth2\Client\Resources\AuthenticatedResourceA
      * @param string $registrantKey
      * @return array
      */
-    public function deleteRegistrant(string $webinarKey, string $registrantKey): ResultSetInterface
+    public function deleteRegistrant(string $webinarKey, string $registrantKey): SimpleResultSet
     {
         $url = $this->getRequestUrl('/organizers/{organizerKey}/webinars/{webinarKey}/registrants/{registrantKey}', [
             'webinarKey' => $webinarKey,